pyproject.toml
[project]
name = "conversion-engine"
version = "0.1.0"
description = "Production-grade SMS + CRM agent for B2B sales development"
requires-python = ">=3.11"
dependencies = [
    # Observability
    "langfuse>=2.0.0",
    "python-dotenv>=1.0.0",
    # tau2-bench runtime deps (so we can import tau2 into our venv)
    "loguru>=0.7.3",
    "rich",
    "tabulate>=0.9.0",
    "fastapi>=0.115.11",
    "uvicorn>=0.34.0",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "docstring-parser>=0.16",
    "litellm>=1.80.15",
    "tenacity>=9.0.0",
    "deepdiff>=8.4.2",
    "addict>=2.4.0",
    "PyYAML>=6.0.2",
    "toml>=0.10.2",
    "typer>=0.12.5",
    "requests>=2.31.0",
    "numpy>=1.26.0",
    "httpx>=0.24.0",
    "openai>=1.0.0",
    "tqdm>=4.66.0",
    "africastalking>=1.2.8",
    "scipy>=1.13.0",
    "hubspot-api-client>=8.0.0",
    "playwright>=1.58.0",
]

[tool.hatch.build.targets.wheel]
packages = []
